refactor: extract audio effects, loop detection, and types from audioManager
Move bus effect chain creation (reverb, filter, compressor, delay) to
audioEffects.ts, waveform loop point detection to loopDetection.ts,
and shared interfaces to audioTypes.ts. AudioManager delegates to
extracted modules, all 125 tests pass unchanged.

1671 lines -> 1387 (manager) + 188 (effects) + 119 (loop) + 31 (types)
